backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0

sqlalchemy==2.0.43

pydantic==2.10.6
pydantic-settings==2.10.1

python-multipart==0.0.20
requests==2.32.4
PyJWT==2.10.1
pwdlib[argon2]==0.3.0
python-dotenv==1.1.1

feedparser==6.0.11

moviepy==1.0.3
Pillow==10.4.0

arabic-reshaper==3.0.0
python-bidi==0.4.2

imageio-ffmpeg==0.6.0
numpy==2.2.6

fal-client==0.7.0
email-validator==2.2.0